modularise slightly
